+++ src/composite.jl
@@ -1,3 +1,4 @@
+
 doc"""
 Allows for combination of several Holiday Calendars.
 """